Add functionality for 5 rounds of game
Implement functionality to play 5 rounds of rock, paper, and scissors.
Previously, the round winner's score was increased within the playRound
function, necessitating two values (humanScore and computerScore) to be
returned at the end of each round to keep track of the updated scores.
In this commit, the source code is also refactored such that instead of
updating scores itself, the playRound function returns the name of the
winner to the playGame function at the end of each round. The playGame
function then updates the score based on the returned value, thus
requiring only a single value (instead of two values) to be returned at
the end of each round to update player scores.