Remove handling of invalid input
Previously, if the user entered an input other than "rock", "paper", or
"scissors", the getHumanChoice function handled it by returning an
"Invalid Choice" string. With the addition of a UI, users select their
choice by clicking on options on the screen, thus making invalid input
handling redundant (the input is always valid, because the user can only
choose from the choices presented by the program).