Refactor function for human and computer choice
Refactor function code used for obtaining both human choice and computer
choice. Instead of using conditional if statement for obtaining the
computer choice, a simpler and shorter switch statement is instead used.
The "humanChoice" argument passed to the getHumanChoice function is the
id of the choice selected by the player, and these ids are all in
lowercase. As such, the conversion of humanChoice to lowercase in the
getHumanChoice function is removed to avoid redundancy (converting an
already lowercase string to lowercase).