Replace event delegation with forEach function
Using event delegation to register event handlers for child elements
caused issues when the mouse was clicked on anywhere besides exactly
the buttons. E.g.: unintentionally clicking on the parent div container
also resulted in an event being fired up; attempting to add images to
represent the choices resulted in an "undefined" player choice as it
selected the image (which does not have a corresponding return value in
the getHumanChoice function) and not the button. This issue is resolved
by using individual event listeners on each button, instead of relying
on event delegation. In this case, regardless of whether an image is
placed inside the button, the event handler will still work due to
the event bubbling from the image to the button; it will also not return
an "undefined" value because the argument for the getHumanChoice
function is a button id (which always has a corresponding return value),
and not event.target.id (which may not have a corresponding value, such
as when the target is an image inside the button element).